ai/requirements.txt
openai
python-dotenv
flask
flask-cors
huggingface_hub
pyarrow